Remove "publishing" restriction for appending sample buffer
In the previous implementation, AppendSampleBuffer() in RTMPStream will not
append sample buffer unless the stream is already publishing. But this will
prevent stream from drawing on view unless the source is the built-in camera,
configured via AttachCamera(). AttachCamera() is able to draw on view when
stream is not publishing, which creates the "preview" experience, as it directly
invokes AppendSampleBuffer() in NetStream.

This commit removes the overridden AppendSampleBuffer() in RTMPStream, which
allows callers to invoke AppendSampleBuffer() in NetStream directly. This will
enable sources other than built-in camera to gain "preview" experience as well.