Update MTHKView.swift
Currently setting `delegate = self` and `enableSetNeedsDisplay = true` is only called on `awakeFromNib()`. This causes the view to never be drawn when creating the view programmatically. 

Added `configureDrawingIfNeded()` to enabled when constructed with frame as well as on `awakeFromNib()`.

Also set `isPaused = true` according to  [Apple's Documentation](https://developer.apple.com/documentation/metalkit/mtkview).